Update exp manager docs and refactor loggers
